+++ Service/Facade/AchatItemService.cs
@@ -5,8 +5,8 @@ namespace DotnetGenerator.Service.Facade;
 
 public interface AchatItemService: IService<AchatItem>
 {
-    public Task<int> DeleteByAchatId(int id);
-    public Task<int> DeleteByProduitId(int id);
-    public Task<List<AchatItem>?> FindByAchatId(int id);
-    public Task<List<AchatItem>?> FindByProduitId(int id);
+    public Task<int> DeleteByAchatId(long id);
+    public Task<int> DeleteByProduitId(long id);
+    public Task<List<AchatItem>?> FindByAchatId(long id);
+    public Task<List<AchatItem>?> FindByProduitId(long id);
 }
